pyproject.toml
[project]
name = "keepomize"
version = "0.1.2"
description = "A filter for Kubernetes manifests that resolves Keeper URIs in Secret resources using ksm"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "ruamel.yaml>=0.17",
]
authors = [
    {name = "Nathan Stien", email = "nathanism@gmail.com"},
]
maintainers = [
    {name = "Nathan Stien", email = "nathanism@gmail.com"},
]
license = "MIT"
keywords = ["kubernetes", "secrets", "keeper", "kustomize"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
"Homepage" = "https://github.com/nathanic/keepomize"
"Bug Tracker" = "https://github.com/nathanic/keepomize/issues"
"Repository" = "https://github.com/nathanic/keepomize"
"Documentation" = "https://github.com/nathanic/keepomize#readme"

[project.scripts]
keepomize = "keepomize.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-p", "no:asyncio",
]


[tool.ruff]
target-version = "py38"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
files = ["src/keepomize"]
mypy_path = "src"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.12.2",
    "mypy>=1.14.1",
]
